refactor: process all validation in single class
